# swagger.yaml
openapi: 3.0.3
info:
  title: Hackathon API
  version: "1.0.0"
  description: |
    Hackathon API — Uploads & attachments (presigned uploads) documentation.
    API routes are relative to /hackathon/v1.
servers:
  - url: http://localhost:3333/hackathon/v1/
    description: Local development server
  - url: https://develop-hackathon-api.224668.xyz/hackathon/v1/
    description: Develop server
  - url: https://main-hackathon-api.224668.xyz/hackathon/v1/
    description: Production server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Attachment:
      type: object
      properties:
        id:
          type: string
          format: uuid
        batch_id:
          type: string
          format: uuid
        uploaded_by:
          type: string
          format: uuid
        kind:
          type: string
        url:
          type: string
          description: Stored object key (not the full presigned URL)
        mime_type:
          type: string
        metadata:
          type: object
        created_at:
          type: string
          format: date-time
paths:
...
  /attachments/{id}:
    get:
      summary: Get attachment metadata and a presigned GET URL
      tags:
        - Attachments
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Attachment metadata + presigned GET URL
          content:
            application/json:
              schema:
                type: object
                properties:
                  attachment:
                    $ref: "#/components/schemas/Attachment"
                  url:
                    type: string
                    description: presigned GET URL
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Attachment not found
        "500":
          description: Server error

    delete:
      summary: Delete an attachment (only uploader allowed)
      tags:
        - Attachments
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: Deleted
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Attachment not found
        "500":
          description: Server error
...